+++ app/routes.py
@@ -3,7 +3,7 @@ import time
 from fastapi import APIRouter
 from pydantic import BaseModel
 
-from nlp.test import get_most_similar_top_ten
+from nlp.fasttext_unit import get_most_similar_top_ten
 from .utils import extract_full_content_with_ocr, extract_foods_with_gpt
 from nlp.cosine_similarity import get_recommendations_with_new_data
