+++ nlp/fasttext_model_trainning.py
